FIX: Tuber and Variety controllers

+++ src/main/java/com/inversionesaraujo/api/business/payload/OrderDataResponse.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class OrderDataResponse {
-    public Integer ship;
-    public Integer pen;
+    public Long ship;
+    public Long pen;
 }
